[Test code] ARM64: Decrease sizes of some values
These values can be changed to smaller types without causing problems.

I've done this in a separate commit because my dev setup doesn't use
ARM. In case something ARM-specific breaks in CI, I'd like to be able to
keep those changes in their own commit so I can debug more easily.

+++ src/vmm/src/arch/aarch64/mod.rs
@@ -129,7 +129,7 @@ mod tests {
         let regions = arch_memory_regions(1usize << 41);
         assert_eq!(1, regions.len());
         assert_eq!(GuestAddress(super::layout::DRAM_MEM_START), regions[0].0);
-        assert_eq!(super::layout::DRAM_MEM_MAX_SIZE, regions[0].1 as u64);
+        assert_eq!(super::layout::DRAM_MEM_MAX_SIZE, regions[0].1);
     }
 
     #[test]
